Add delegate to AutoCompleteTextField
Additionally, I prefixed the `dataSource` and `delegate` properties with
`autoComplete` to differentiate it from the inherited `delegate`
property from `UITextField`.

Added a routine which, when given the suggested text, returns the raw
form of the suggestion. This ensures the data is consistent. For
example, if this were a Core Data application, multiple spellings of the
same intended value will create multiple records, which would confuse
the user.